Initial Commit of k means report and code
Project Description: We will be researching data science and machine learning techniques that can be used to fine tune an Intrusion Detection System (IDS). The goal of this research is to create a working prototype to aid in the reduction of customizations organizations need to make when implementing and maintaining a working and effective IDS. This prototype could then be further enhanced by future research teams.